python-core/pyproject.toml
[project]
name = "python-core"
version = "0.2.0"
dependencies = [
  "aiohttp>=3.11.18",
  "azure-ai-documentintelligence==1.0.2",
  "azure-ai-projects",
  "azure-ai-textanalytics",
  "azure-cosmos>=4.9.0",
  "azure-identity",
  "azure-mgmt-cognitiveservices",
  "azure-search-documents==11.5.2",
  "azure-storage-blob",
  "black>=24.3.0",
  "certifi>=2022.12.07",
  "docopt",
  "duckdb",
  "Faker",
  "h11>=0.14.0",
  "httpx",
  "huggingface_hub",
  "Jinja2>= 3.1.4",
  "openai>=1.78.1",
  "promptflow-tools",
  "promptflow",
  "psutil",
  "pymongo>=4.6.3",
  "pypistats",
  "pytest-asyncio>=1.0.0",
  "pytest-cov>=6.1.1",
  "pytest>= 8.3.5",
  "python-dotenv",
  "pytz",
  "redis",
  "semantic-kernel",
  "six==1.17.0",
  "tiktoken"
]
requires-python = ">=3.13"
authors = [
  {name = "Chris Joakim", email = "christopher.joakim@gmail.com"}
]
maintainers = [
  {name = "Chris Joakim", email = "christopher.joakim@gmail.com"}
]
description = "A core set of python code used elsewhere in this repo"
readme = "readme.md"
license = "MIT"
license-files = ["LICEN[CS]E.*"]
keywords = ["python", "pyproject.toml", "uv", "ruff"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python"
]

[project.urls]
Homepage = "https://github.com/cjoakim/azure-ai-explorations/"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
ignore = ["E722"]
